refactor-settings * Move services/firebase.ts to settings/firebase.ts
