+++ src/containers/Admin/ShareTab/components/BoardOwnersList.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { Table, TableHead, TableRow, HeaderCell, TableBody } from '@entur/table'
 import { useUser } from '../../../../UserProvider'
 import type { BoardOwnersData, Invite } from '../../../../types'
-import { removeSentBoardInviteAsOwner } from '../../../../services/firebase'
+import { removeSentBoardInviteAsOwner } from '../../../../settings/firebase'
 import { removeFromOwners } from '../../../../settings/FirestoreStorage'
 import { SharedWithRows } from './SharedWithRows'
