+++ src/containers/LandingPage/LandingPage.tsx
@@ -6,7 +6,7 @@ import { Contrast } from '@entur/layout'
 import { ForwardIcon } from '@entur/icons'
 import { GridContainer, GridItem } from '@entur/grid'
 import { DEFAULT_SETTINGS } from '../../settings/settings'
-import { createSettings } from '../../services/firebase'
+import { createSettings } from '../../settings/firebase'
 import { SearchPanel } from './SearchPanel/SearchPanel'
 import { TypographyCarousel } from './TypographyCarousel/TypographyCarousel'
 import './LandingPage.scss'
